Update: Fan shroud with a bit more clearance
The fan shroud in CAD seems to be a newer version as the one from the original repo, so I've uploaded the newer version (the mounting holes for the shroud have a sacrificial layer so that the holes can be printed clean). Additionally I've added a bit more clearance for the hotend fan as most of my fans are a bit thicker than 10mm. Also added 0,1 mm more clearance to the top of the right blower fan. Now the clearance is the same left and right.